refactor(keyman): clear detekt on the new module, and stop losing the reason
:core:keyman is clean. Seven findings, none baselined: the baseline is at zero
and adding to it for new code would defeat the point of having re-zeroed it.

Most were real simplifications. Two `when` blocks over closed types had an
`else` that would have swallowed a new case silently; the sp check is now a set
membership, which is what it always meant. The rule prefilter had a `when`
whose branches differed only in which value they compared, so it is one
comparison and a named variable.

Two suppressions stay, each with its reason in the code. KmxParser.parse has
sixteen exits and each is a distinct way the bytes can be wrong; folding them
together would hide which check failed, which is the whole job of that
function. processGroup's jumps are the reference implementation's own control
flow, and restructuring them would make the two unreadable against each other,
which is the only way to tell whether the port is still right.

Dropped a dead branch in process(): `!ok && emitKeystroke` is subsumed by the
line below it, and `ok` was never otherwise consulted. A group that matched
nothing declines only when it asked to, which emitKeystroke already records.

The download failure no longer swallows its exception. The reason is carried
on Outcome.Failed and shown on the row, worded through
ToolHttp.friendlyMessage the way the addon downloader words its own failures,
so "Download failed" becomes something the user can act on. Cancellation is
caught separately: it means the user left the screen, and treating it as a
failure would report one that never happened.

Two pre-existing findings in AddonPreviewContent.kt are left alone. They
predate this work, :feature:addons detekt is not part of the staticAnalysis
gate, and fixing a file this change never touched is not this commit's job.

+++ core/keyman/src/test/java/com/wasimaster/wmkeyboard/core/keyman/KeymanPackageTest.kt
@@ -77,13 +77,13 @@ class KeymanPackageTest {
     /** Reading stops at the cap rather than at whatever the header claims. */
     @Test
     fun `an oversized entry is refused`() {
-        val huge = ByteArray((KeymanLimits.MAX_KMX_BYTES + 1024).toInt())
+        val huge = ByteArray(KeymanLimits.MAX_KMX_BYTES + 1024)
         assertNull(rules(zip("big.kmx" to huge), "big"))
     }
 
     @Test
     fun `an entry at the cap is still read`() {
-        val atCap = ByteArray(KeymanLimits.MAX_KMX_BYTES.toInt()) { 3 }
+        val atCap = ByteArray(KeymanLimits.MAX_KMX_BYTES) { 3 }
         val got = rules(zip("ok.kmx" to atCap), "ok")
         assertEquals(atCap.size, got?.size)
     }
